requirements.txt
# Copyright (c) 2024, ETH Zurich
# python~=3.11.9
numpy~=1.24
scipy~=1.10
matplotlib~=3.7
ruamel.yaml~=0.17
./nist_lookup
pandas~=1.5.3
seaborn~=0.12.2
h5py~=3.9
tqdm~=4.66

# nist_lookup doesn't specify the required SQLAlchemy version but is not compatible with 1.x
SQLAlchemy~=1.4

# dev dependencies
mypy~=1.3
maturin~=1.2
freezegun==1.2.2
jupyter~=1.0.0

# only used for the performance testing script
chevron==0.14.0
types-chevron==0.14.2.4

# using this for matplotlib type stubs. todo: remove once matblotlib 3.8 is released
data-science-types==0.2.23
